pyproject.toml
[project]
name = "crotal"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "python-jose",
    "itsdangerous",
    "pydantic-settings",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"